Fix toolbar for task item
