+++ src/components/list-detail/task-item/actions/style.module.scss
@@ -1,0 +1,39 @@
+.actions :global {
+  @apply flex items-center gap-x-2.5;
+  .tool-desktop {
+    @apply hidden h-6 sm:block;
+    span {
+      @apply hidden;
+    }
+  }
+  .status {
+    @apply relative flex h-full items-center;
+    .MuiOutlinedInput-root {
+      @apply h-5 w-5 rounded-full lg:h-8 lg:w-auto lg:rounded;
+    }
+
+    .MuiSelect-select {
+      @apply h-full w-full;
+      padding: 0 !important;
+      .mobile-icon {
+        @apply absolute flex h-[23px] w-full items-center justify-center lg:hidden;
+      }
+    }
+
+    .MuiSelect-select[aria-expanded='true'] {
+      .mobile-icon {
+        transform: rotate(180deg) !important;
+      }
+    }
+    .status-name {
+      @apply hidden pl-3 pr-8 lg:block;
+    }
+    .MuiOutlinedInput-notchedOutline {
+      @apply border-0;
+    }
+
+    .MuiSelect-icon {
+      @apply hidden fill-white lg:block;
+    }
+  }
+}
