+++ src/components/list-detail/task-item/index.tsx
@@ -1,26 +1,17 @@
 import {useSortable} from '@dnd-kit/sortable';
 import {CSS} from '@dnd-kit/utilities';
-import {SelectChangeEvent} from '@mui/material';
 import classNames from 'classnames';
-import {useRouter} from 'next/router';
 
-import {ROUTES} from '@/configs/routes.config';
 import Checkbox from '@/core-ui/checkbox';
-import Icon from '@/core-ui/icon';
-import IconButton from '@/core-ui/icon-button';
 import api from '@/data/api/index';
 import {IStatus} from '@/data/api/types/list.type';
 import {ITaskResponse} from '@/data/api/types/task.type';
 import {socketUpdateList} from '@/data/socket';
-import {MUI_ICON} from '@/utils/mui-icon';
 
-import Status from '../../common/statusSelect';
-import Tool, {IToolProps} from '../toolbar/tool';
-import ToolMenu from '../toolbar/tool-menu';
-// import style from './style.module.scss';
+import Actions from './actions';
 import style from './style.module.scss';
 
-interface IProp {
+export interface ITaskItemProps {
   task: ITaskResponse;
   onEdit?: () => void;
   onDelete?: () => void;
@@ -29,11 +20,10 @@ interface IProp {
   readonly: boolean;
 }
 
-export default function TaskItem({task, onEdit, onDelete, statusList, isSelect, readonly}: IProp) {
-  const router = useRouter();
+export default function TaskItem(props: ITaskItemProps) {
+  const {task, statusList, isSelect, readonly} = props;
   const {attributes, listeners, setNodeRef, transform, transition, isDragging} = useSortable({id: task!.id!});
   statusList?.sort((a, b) => a.id - b.id);
-  const statusValue = statusList.filter(e => e.id === task.statusId)[0];
 
   const setDone = (id: string, isDone: boolean) => {
     if (!id) return;
@@ -48,36 +38,6 @@ export default function TaskItem({task, onEdit, onDelete, statusList, isSelect, 
     transition,
     opacity: isDragging ? 0.5 : 1
   };
-  const onChangeStatus = (event: SelectChangeEvent<unknown>) => {
-    api.task
-      .update({id: task.id, statusId: Number(event.target.value)})
-      .then(socketUpdateList)
-      .catch(() => {});
-  };
-  const onDetail = (taskId: string) => {
-    router.push(ROUTES.TASK + '/' + taskId);
-  };
-
-  const deleteToolProps: IToolProps = {
-    icon: <Icon name="ico-trash-2" />,
-    text: 'Delete',
-    hidden: false,
-    onClick: onDelete
-  };
-  const editToolProps: IToolProps = {
-    icon: <Icon name="ico-edit" />,
-    text: 'Edit',
-    onClick: onEdit
-  };
-  const goDetailToolProps: IToolProps = {
-    icon: <Icon name="ico-chevron-right" />,
-    text: 'Detail',
-    onClick: () => onDetail(task.id)
-  };
-
-  const toolMenuItems = [deleteToolProps, editToolProps, goDetailToolProps]
-    .filter(item => !item.hidden)
-    .map((item, idx) => <Tool key={idx} {...{...item, className: 'flex-row-reverse'}} />);
 
   return (
     <div
@@ -98,23 +58,11 @@ export default function TaskItem({task, onEdit, onDelete, statusList, isSelect, 
         }
       }}
     >
-      <Checkbox checked={task?.isDone} onChange={() => setDone(task!.id!, task!.isDone)} disabled={readonly} />
-      <p className={`h6 ${task!.isDone ? 'checked' : ''}`} onClick={() => setDone(task!.id!, task!.isDone)}>
-        {`${task!.name}`}
+      <Checkbox checked={task?.isDone} onChange={() => setDone(task.id, task.isDone)} disabled={readonly} />
+      <p className={`h6 ${task.isDone ? 'checked' : ''}`} onClick={() => setDone(task.id, task.isDone)}>
+        {task.name}
       </p>
-      <div className="actions">
-        {statusList && <Status className={style.status} items={statusList} readOnly={readonly} status={statusValue} onChange={e => onChangeStatus(e)} />}
-        {!readonly && (
-          <>
-            <IconButton name="ico-edit" className="hidden sm:block" onClick={onEdit} />
-            <IconButton name="ico-trash-2" className="hidden sm:block" onClick={onDelete} />
-            <div className="action-mobile">
-              <ToolMenu icon={<MUI_ICON.MORE_HORIZON />} items={toolMenuItems} />
-            </div>
-          </>
-        )}
-        <IconButton name="ico-chevron-right" className="hidden sm:block" onClick={() => onDetail(task.id)} />
-      </div>
+      <Actions {...props} />
     </div>
   );
 }
