+++ src/components/list-detail/task-item/style.module.scss
@@ -1,49 +1,11 @@
 .task :global {
   @apply relative flex min-h-[48px] w-full justify-between bg-white;
-  @apply border border-t-0;
-  @apply pl-4 pr-2;
+  @apply border border-t-0 pl-4 pr-2;
   .h6 {
     @apply mx-3 flex w-full cursor-pointer items-start py-3 leading-5;
     word-break: break-word;
   }
   .checked {
     @apply line-through opacity-50;
-  }
-  .actions {
-    @apply flex gap-x-3;
-    .action-mobile {
-      @apply py-3 pr-2;
-      @apply flex items-center sm:hidden;
-    }
-  }
-}
-.status :global {
-  @apply relative flex h-full items-center;
-  .MuiOutlinedInput-root {
-    @apply h-5 w-5 rounded-full lg:h-8 lg:w-auto lg:rounded;
-  }
-
-  .MuiSelect-select {
-    @apply h-full w-full;
-    padding: 0 !important;
-    .mobile-icon {
-      @apply absolute flex h-[23px] w-full items-center justify-center lg:hidden;
-    }
-  }
-
-  .MuiSelect-select[aria-expanded='true'] {
-    .mobile-icon {
-      transform: rotate(180deg) !important;
-    }
-  }
-  .status-name {
-    @apply hidden pl-3 pr-8 lg:block;
-  }
-  .MuiOutlinedInput-notchedOutline {
-    @apply border-0;
-  }
-
-  .MuiSelect-icon {
-    @apply hidden fill-white lg:block;
   }
 }
